# src/docs/openapi.generated.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API RRHH Hospital - Personal",
    "version": "1.0.0",
    "description": "OpenAPI enterprise (auto)"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Cc": {
        "type": "object",
        "properties": {
          "dni": {
            "type": "string"
          },
          "observaciones": {
            "type": "string"
          },
          "fecha": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
...
      "Jefaturas": {
        "type": "object",
        "properties": {
          "sector": {
            "type": "string"
          },
          "jefe": {
            "type": "string"
          }
        }
      },
...
      "Ley": {
        "type": "object",
        "properties": {
          "ley_nombre": {
            "type": "string"
          },
          "codigoexp": {
            "type": "string"
          },
          "leyactiva": {
            "type": "string"
          }
        }
      },
...
      "Ocupacion": {
        "type": "object",
        "properties": {
          "nombre_ocupacion": {
            "type": "string"
          },
          "codigo": {
            "type": "string"
          },
          "regimen_horario_id": {
            "type": "string"
          },
          "es_insalubre": {
            "type": "string"
          },
          "especialidad_id": {
            "type": "string"
          },
          "ley_id": {
            "type": "string"
          },
          "agrupamiento": {
            "type": "string"
          },
          "grado": {
            "type": "string"
          },
          "asignacion": {
            "type": "string"
          },
          "activo": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/cc": {
      "get": {
        "tags": [
          "cc"
        ],
        "summary": "Listar cc",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "id"
            }
          },
          {
            "name": "order",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ],
              "default": "ASC"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "cc"
        ],
        "summary": "Crear cc",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Cc"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Creado"
          }
        }
      }
    },
    "/api/cc/{id}": {
      "get": {
        "tags": [
          "cc"
        ],
        "summary": "Obtener cc",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "No encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "cc"
        ],
        "summary": "Actualizar cc",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Cc"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "No encontrado"
          }
        }
      },
      "delete": {
        "tags": [
          "cc"
        ],
        "summary": "Eliminar cc (soft-delete)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Borrado"
          },
          "404": {
            "description": "No encontrado"
          }
        }
      }
    },
...
    "/api/jefaturas": {
      "get": {
        "tags": [
          "jefaturas"
        ],
        "summary": "Listar jefaturas",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "id"
            }
          },
          {
            "name": "order",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ],
              "default": "ASC"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "jefaturas"
        ],
        "summary": "Crear jefaturas",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Jefaturas"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Creado"
          }
        }
      }
    },
    "/api/jefaturas/{id}": {
      "get": {
        "tags": [
          "jefaturas"
        ],
        "summary": "Obtener jefaturas",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "No encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "jefaturas"
        ],
        "summary": "Actualizar jefaturas",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Jefaturas"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "No encontrado"
          }
        }
      },
      "delete": {
        "tags": [
          "jefaturas"
        ],
        "summary": "Eliminar jefaturas (soft-delete)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Borrado"
          },
          "404": {
            "description": "No encontrado"
          }
        }
      }
    },
...
    "/api/ley": {
      "get": {
        "tags": [
          "ley"
        ],
        "summary": "Listar ley",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "id"
            }
          },
          {
            "name": "order",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ],
              "default": "ASC"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "ley"
        ],
        "summary": "Crear ley",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Ley"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Creado"
          }
        }
      }
    },
    "/api/ley/{id}": {
      "get": {
        "tags": [
          "ley"
        ],
        "summary": "Obtener ley",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "No encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "ley"
        ],
        "summary": "Actualizar ley",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Ley"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "No encontrado"
          }
        }
      },
      "delete": {
        "tags": [
          "ley"
        ],
        "summary": "Eliminar ley (soft-delete)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Borrado"
          },
          "404": {
            "description": "No encontrado"
          }
        }
      }
    },
...
    "/api/ocupacion": {
      "get": {
        "tags": [
          "ocupacion"
        ],
        "summary": "Listar ocupacion",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "id"
            }
          },
          {
            "name": "order",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ],
              "default": "ASC"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "ocupacion"
        ],
        "summary": "Crear ocupacion",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Ocupacion"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Creado"
          }
        }
      }
    },
    "/api/ocupacion/{id}": {
      "get": {
        "tags": [
          "ocupacion"
        ],
        "summary": "Obtener ocupacion",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "No encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "ocupacion"
        ],
        "summary": "Actualizar ocupacion",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Ocupacion"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "No encontrado"
          }
        }
      },
      "delete": {
        "tags": [
          "ocupacion"
        ],
        "summary": "Eliminar ocupacion (soft-delete)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Borrado"
          },
          "404": {
            "description": "No encontrado"
          }
        }
      }
    },
...
  }
}